requirements.txt
asgiref==3.4.1
coverage==6.1.1
Django==3.2.9
django-extensions==3.1.3
djangorestframework==3.12.4
psycopg2-binary==2.9.1
pytz==2021.3
sqlparse==0.4.2
